StudentUserController

+++ Controller/Student_user-controller.js
@@ -10,28 +10,19 @@ const OTP = require("../Model/otp");
 const jwt = require("jsonwebtoken");
 const { hashData, verifyHashedData } = require("../Middleware/hashDataHandler");
 const createToken = require("../Middleware/createToken");
-const {
-  verifyStudentEmailwithOTP,
-} = require("../Controller/verifyemailController");
+const {verifyStudentEmailwithOTP} = require("../Controller/verifyemailController");
 const catchAsync = require("../utils/catchAsync");
 const AppError = require("../utils/appError");
 
-const getAllStudent_user = async (req, res, next) => {
-  let getStudent_users;
-  try {
-    getStudent_users = await Student_user.find();
-  } catch (err) {
-    next(err);
-  }
+const getAllStudent_user =catchAsync( async (req, res, next) => {
+  const getStudent_users = await Student_user.find();
   if (!getStudent_users) {
-    return res.status(500).json({
-      message: "Internal Server Error",
-    });
+    return  next(new AppError('not found Student_user', 404));
   }
-  return res.status(200).json({
+   res.status(200).json({
     Student_user: getStudent_users,
   });
-};
+});
 const studentSignUp = catchAsync(async (req, res, next) => {
   let {
     name,
@@ -43,22 +34,15 @@ const studentSignUp = catchAsync(async (req, res, next) => {
     city,
     dob,
   } = req.body;
-  var emailFormat = /^[A-Z0-9+_.-]+@[A-Z0-9.-]+$/;
-  if (email.match(emailFormat)) {
-    return next(new AppError("Email not matched", 400));
-  }
   if (
     !(name || email || password || phone1 || phone2 || state || city || dob)
   ) {
-    return res.status(500).json({
-      message: "Internal server Error",
-    });
+    return  next(new AppError('somting wrong please verify  values', 400));
   }
   const userAvailable = await Student_user.findOne({ email });
   if (userAvailable) {
-    return res.status(409).json({
-      message: "email already Exist",
-    });
+    return  next(new AppError('email already Exist', 409));
+  
   }
   const hashPassword = await hashData(password);
   const user = await Student_user.create({
@@ -72,17 +56,13 @@ const studentSignUp = catchAsync(async (req, res, next) => {
     dob,
   });
   if (!user) {
-    return res.status(500).json({
-      message: "user Error",
-    });
+    return  next(new AppError('not found', 404));
   }
   await verifyStudentEmailwithOTP({ email });
-
   return res.status(201).json({
     Student_user: user,
   });
 });
-
 const Student_userLogin =catchAsync( async (req, res, next) => {
     const { email, password } = await req.body;
     const userFetch = await Student_user.findOne({ email });
@@ -98,17 +78,13 @@ const Student_userLogin =catchAsync( async (req, res, next) => {
       const tokeData = { userId: userFetch._id, email };
       const token = await createToken(tokeData);
       userFetch.token = token;
-      console.log(userFetch);
       res.status(200).json({
         userFetch,
         image,
       });
     }
-  })
-;
-
-const updateStudent_user = async (req, res, next) => {
-  try {
+});
+const updateStudent_user =catchAsync( async (req, res, next) => {
     const userId = req.params.id;
     let { name, email, phone1, phone2, state, city, dob } = req.body;
     const updateFields = { name, email, phone1, phone2, state, city, dob };
@@ -132,9 +108,7 @@ const updateStudent_user = async (req, res, next) => {
       new: true,
     });
     if (!user) {
-      return res.status(500).json({
-        message: "User not found",
-      });
+      return  next(new AppError('user Not Available in DB', 400));
     }
     if (newImage) {
       res.status(200).json({
@@ -147,67 +121,37 @@ const updateStudent_user = async (req, res, next) => {
         image: oldImage,
       });
     }
-  } catch (err) {
-    res.status(400).json({
-      err: err.message,
-    });
-  }
-};
-
-const currentUser = async (req, res, next) => {
-  return res.status(200).json({
-    res: req.user,
-  });
-};
-
-const deleteStudent_User = async (req, res, next) => {
+});
+const deleteStudent_User =catchAsync( async (req, res, next) => {
   const id = req.params.id;
-  let user;
-  try {
-    user = await Student_user.findByIdAndRemove(id);
-    console.log(user);
-  } catch (err) {
-    next(err);
-  }
-  if (!user) {
-    return res.status(500).json({
-      message: "Internal Server Error",
-    });
+  const user = await Student_user.findByIdAndDelete(id);
+  if(!user){
+    return  next(new AppError('user Not Available in DB', 404));
   }
   return res.status(200).json({
     message: "Deleted Succcess",
   });
-};
-const getStudet_user = async (req, res, next) => {
+});
+const getStudet_user =catchAsync( async (req, res, next) => {
   const id = req.params.id;
-  let user;
-  try {
-    user = await Student_user.findById(id);
-    console.log(user);
-  } catch (err) {
-    next(err);
-  }
+  const user = await Student_user.findById(id);
   if (!user) {
-    return res.status(500).json({
-      message: "Internal Server Error",
-    });
+    return  next(new AppError('user Not Available in DB', 404));
   }
   return res.status(200).json({
     user: user,
   });
-};
-
-const authUser = (req, res) => {
-  res
-    .status(200)
-    .send(`you are in the private route  ${req.currentUser.email}`);
-};
-
-const uploadImg = async (req, res) => {
-  try {
+});
+const uploadImg =catchAsync( async (req, res) => {
     const imagepath = req.file?.path;
     const studentUser = await Student_user.findById(req.params.id);
+    if(!studentUser){
+      return  next(new AppError('user Not Available in DB', 404));
+    }
     const oldImage = await StudentUserImg.findOne({ user: studentUser._id });
+    if(!oldImage){
+      return  next(new AppError('Not StudentUserImg in DB', 404));
+    }
     if (oldImage) {
       await StudentUserImg.deleteOne({ _id: oldImage._id });
     }
@@ -221,18 +165,12 @@ const uploadImg = async (req, res) => {
       studentUser,
       newImage,
     });
-  } catch (err) {
-    console.error(err);
-    res.status(500).send("An error occurred while uploading the file.");
-  }
-};
+});
 
 exports.getAllStudent_user = getAllStudent_user;
 exports.studentSignUp = studentSignUp;
 exports.updateStudent_user = updateStudent_user;
 exports.deleteStudent_User = deleteStudent_User;
 exports.getStudet_user = getStudet_user;
-exports.currentUser = currentUser;
 exports.Student_userLogin = Student_userLogin;
-exports.authUser = authUser;
 exports.uploadImg = uploadImg;
